pyproject.toml
[tool.poetry]
name = "social_network_api"
version = "0.1.0"
description = "Social network API"
authors = ["Artemooon <artem.logachov773@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
PyJWT = "^2.1.0"
psycopg2-binary = "^2.9.1"
django-redis = "^5.0.0"
drf-yasg = "^1.20.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
